Moves BlitManager to its own file
