+++ python_scripts/live_motor_plotter.py
@@ -1,13 +1,13 @@
-from xml.etree.ElementTree import PI
 import lcm
-from matplotlib import animation
 from lcm_types.ManyMotorLog import ManyMotorLog
 import matplotlib.pyplot as plt
 import numpy as np
-np.set_printoptions(suppress=True)
-from matplotlib.animation import FuncAnimation
 import collections
 import time
+from blit_manager import BlitManager
+
+# NumPy Options
+np.set_printoptions(suppress=True)
 
 # Constants
 PLIM = np.pi
@@ -16,7 +16,6 @@ TLIM = 0.01
 WINDOWLENGTH = 2000
 PLOTPERIOD = 1000
 TIMEOUT = 500
-
 
 # Queues
 timestamps = collections.deque(np.zeros(WINDOWLENGTH))
@@ -27,80 +26,6 @@ velocities = collections.deque(np.zeros((WINDOWLENGTH,13)))
 torques = collections.deque(np.zeros((WINDOWLENGTH,13)))
 velo_filt = collections.deque(np.zeros((WINDOWLENGTH,13)))
 
-class BlitManager:
-    def __init__(self, canvas, animated_artists=()):
-        """
-        Parameters
-        ----------
-        canvas : FigureCanvasAgg
-            The canvas to work with, this only works for sub-classes of the Agg
-            canvas which have the `~FigureCanvasAgg.copy_from_bbox` and
-            `~FigureCanvasAgg.restore_region` methods.
-
-        animated_artists : Iterable[Artist]
-            List of the artists to manage
-        """
-
-        self.canvas = canvas
-        self._bg = None
-        self._artists = []
-
-        for a in animated_artists:
-            self.add_artist(a)
-        # grab the background on every draw
-        self.cid = canvas.mpl_connect("draw_event", self.on_draw)
-
-    def on_draw(self, event):
-        """Callback to register with 'draw_event'."""
-        cv = self.canvas
-        if event is not None:
-            if event.canvas != cv:
-                raise RuntimeError
-        self._bg = cv.copy_from_bbox(cv.figure.bbox)
-        self._draw_animated()
-
-    def add_artist(self, art):
-        """
-        Add an artist to be managed.
-
-        Parameters
-        ----------
-        art : Artist
-
-            The artist to be added.  Will be set to 'animated' (just
-            to be safe).  *art* must be in the figure associated with
-            the canvas this class is managing.
-
-        """
-        if art.figure != self.canvas.figure:
-            raise RuntimeError
-        art.set_animated(True)
-        self._artists.append(art)
-
-    def _draw_animated(self):
-        """Draw all of the animated artists."""
-        fig = self.canvas.figure
-        for a in self._artists:
-            fig.draw_artist(a)
-
-    def update(self):
-        """Update the screen with animated artists."""
-        cv = self.canvas
-        fig = cv.figure
-        # paranoia in case we missed the draw event,
-        if self._bg is None:
-            self.on_draw(None)
-        else:
-            # restore the background
-            cv.restore_region(self._bg)
-            # draw all of the animated artists
-            self._draw_animated()
-            # update the GUI state
-            cv.blit(fig.bbox)
-        # let the GUI event loop process anything it has to do
-        cv.flush_events()
-
-    
 
 def init_plot():
 
@@ -115,7 +40,7 @@ def init_plot():
     axs[4].set_ylim(-PLIM, PLIM)
     axs[5].set_ylabel("dt")
     axs[5].set_ylim(0, TLIM)
-    
+
     axs[6].set_title("Velocities")
     for i in range(4):
         axs[6 + i].set_ylim(-VLIM, VLIM) # Leg joint velocities limits
@@ -155,13 +80,13 @@ def update_plot_lines(lns):
 
     for ln in lns:
         ln.set_xdata(times)
-    
+
     for i in range(13):
         lns[i].set_ydata(pos[:, i])
 
     i += 1
     lns[i].set_ydata(np.concatenate(([0],np.diff(times))))
-    
+
     i += 1
     for i, j in zip(range(i, i + 12), range(12)):
         lns[i].set_ydata(vel[:, j])
@@ -170,7 +95,7 @@ def update_plot_lines(lns):
     max_x = np.max(times)
     for ax in axs:
         ax.set_xlim(min_x,max_x)
-    
+
 
 def my_handler(channel, data):
 
@@ -181,7 +106,7 @@ def my_handler(channel, data):
     margins.popleft()
     torques.popleft()
     message_duration.popleft()
-    
+
     timestamps.append(msg.timestamp)
     velocities.append(msg.velocities)
     positions.append(msg.positions)
@@ -211,7 +136,7 @@ if __name__ == '__main__':
             count += 1
             currTime = int(time.clock_gettime_ns(time.CLOCK_MONOTONIC_RAW)/1e6) #millis
             if currTime - lastTime > PLOTPERIOD:
-                
+
                 update_plot_lines(lns)
 
                 bm.update()
@@ -219,7 +144,7 @@ if __name__ == '__main__':
                 plt.draw()
                 plt.pause(.0001)
                 lastTime = currTime
-            
+
 
     except KeyboardInterrupt:
         pass
